chore: burn down nine punchlist rows and retire cleanup-punchlist.md (#3540)
* chore: burn down nine punchlist rows and retire cleanup-punchlist.md

Nine mechanical Targeted PR / Inline rows fixed:

- TMD-1: the tokio-mutex lint now recurses into all of src/ instead of
  scanning only top-level files. The recursive scan immediately found
  six real violations in notebook_sync_server/tests.rs (room RwLock
  guards held across the test loader's awaits) — the test loader is
  split into async prepare_notebook_load (file/blob work, no doc) +
  sync apply_notebook_load (doc mutation inside the lock block) and
  all six sites migrated. Zero violations across the full tree.
- WP-6: known frame types dispatch to an exhaustive
  typed_frame_size_limits(NotebookFrameType) with no wildcard arm; a
  new variant fails to compile until it gets an explicit cap.
- WP-7: PROTOCOL_VERSION / MIN_PROTOCOL_VERSION are u8 — the wire's
  actual width — with explicit u32::from() widening at JSON sites.
- WP-11: recv_typed_frame classifies the type byte before allocating;
  unknown forward-compat frames skip via bounded 8 KiB discard reads
  instead of a body-sized allocation up to the 100 MiB ceiling.
- EP-10: the runtime agent select! is biased with the lifecycle arm
  ahead of bounded work, closing the window where work could be
  polled before lifecycle drain.
- EP-11: the empty-flushes fast path is gone — every flush_then_signal
  rides the priority committer queue, so a no-output execution's
  ExecutionDone can't jump an earlier execution's queued stream
  commit. Test updated to pin the new ordering contract.
- MSL-1: ApproveProjectEnvironment broadcasts and re-evaluates trust
  immediately after the allowlist write, like the other approval
  entry points.
- BS-3: MultipartUploadState spawns a TTL-paced periodic sweep
  (Weak-held, exits when the peer state drops) so an idle daemon no
  longer retains expired staging dirs forever.
- FSB-3: the stableDomOrder comment records why code-unit sort is
  correct (stable deterministic total order, identical on every
  client; locale-aware collation would not be).

The punchlist itself is retired: all 21 remaining open rows migrated
into "Tracked follow-ups" sections in their owning ADRs
(typed-frame-v4-wire-protocol, document-split, execution-pipeline,
blob-storage-and-content-addressing, hosted-room-authorization,
hosted-output-origin-isolation, kernel-env-trust,
captured-environment-lifecycle, mcp-session-lifecycle,
tokio-mutex-discipline), every cross-reference updated, and the ADR
README convention now says to track smells in the owning ADR.


* ci: retrigger after runner disk-space failure


---------

+++ crates/notebook-sync/src/connect.rs
@@ -782,7 +782,7 @@ fn parse_typed_bootstrap_payload(payload: &[u8]) -> Result<ConnectionBootstrap, 
 /// version differences for debugging (e.g., a daemon rebuilt from a
 /// different commit).
 fn check_daemon_protocol_version(caps: &ProtocolCapabilities) {
-    let expected = notebook_protocol::connection::PROTOCOL_VERSION;
+    let expected = u32::from(notebook_protocol::connection::PROTOCOL_VERSION);
 
     if let Some(remote) = caps.protocol_version {
         if remote != expected {
